requirements.txt
sib_api_v3_sdk==7.6.0
python-dotenv==1.0.0